task 0

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,6 +1,5 @@
-#include <stdio.h>
-#include <stdlib.h>
 #include "lists.h"
+
 /**
 * print_listint - print all the elements of a list
 * @h: listint_t
@@ -8,7 +7,7 @@
 */
 size_t print_listint(const listint_t *h)
 {
-int i = 0;
+size_t i = 0;
 while (h)
 {
 printf("%d\n", h->n);
